gas optimizations: transient reentrancy guard, config packing, mint-path caching
- Swap ReentrancyGuard for ReentrancyGuardTransient on both contracts
  (OZ 5.6.1, Cancun TSTORE) — ~2.1k gas per nonReentrant call.
- Pack ContractConfig from 4 slots to 3 by narrowing platform/staking fees
  from uint128 to uint64 (MAX_*_FEE = 1_000_000 fits with plenty of headroom).
- Cache config once at the top of mintSong and thread through _processMint,
  _distributePayments, and _creditShell to avoid re-loading 3 slots per mint.
- Optimize the split-distribution loop: cache length, use storage pointer
  for SplitRecipient, unchecked increment.
- Mark name/symbol/claimRewards external (never called internally).

+++ script/Deploy.s.sol
@@ -9,9 +9,9 @@ import {Config} from "./helpers/Config.s.sol";
 contract DeployTortoise is Config {
     function run() public {
         uint256 deployerPrivateKey = vm.envUint("DEPLOYER_PRIVATE_KEY");
-        uint128 platformFee = uint128(vm.envUint("INITIAL_PLATFORM_FEE"));
+        uint64 platformFee = uint64(vm.envUint("INITIAL_PLATFORM_FEE"));
         uint128 defaultPrice = uint128(vm.envUint("INITIAL_SONG_PRICE"));
-        uint128 stakingFee = uint128(vm.envUint("INITIAL_STAKING_FEE"));
+        uint64 stakingFee = uint64(vm.envUint("INITIAL_STAKING_FEE"));
 
         address usdcAddress = getUsdcAddress();
         address tortAddress = getTortAddress();
